# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Influence Hub Backend API Documentation",
      "description": "Back-end API instrunctions to Influence Hub application",
      "contact": {
        "name": "Rafael Ribeiro",
        "email": "rafittu@outlook.com"
      },
      "version": "1.0.0"
    },
    "paths": {
...
        "/influencer/create": {
            "post": {
                "summary": "Create influencer user",
                "description": "Route to register new influencer.",
                "tags": ["Influencer"],
                "security": [
                    {
                    "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Bearer {accesToken}",
                    "schema": {
                        "type": "string"
                    }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {                    
                            "schema": {
                                "$ref": "#/components/schemas/CreateInfluencer"
                            },
                            "example": {
                                "name": "Ygona Moura",
                                "username": "ygona.moura",
                                "email": "ygona@email.com",
                                "niches": ["Artes", "Beleza & Maquiagem"],
                                "reach": 125000,
                                "photo": "https://example.photo.aws.com",
                                "zipCode": "12345678",
                                "street": "Ipê",
                                "number": "45"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Influencer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "$ref": "#/components/schemas/InfluencerInterface"
                                },
                                "example": {
                                "id": "1",
                                "name": "Ygona Moura",
                                "username": "ygona.moura",
                                "email": "ygona@email.com",
                                "niches": ["Artes", "Beleza & Maquiagem"],
                                "reach": "125000",
                                "photo": "https://example.photo.aws.com",
                                "address": {
                                    "street": "Ipê",
                                    "number": "45",
                                    "zipCode": "12345678",
                                    "city": "Belo Horizonte",
                                    "state": "MG"
                                },
                                "createdAt": "2024-09-05T03:29:37.743Z",
                                "updatedAt": "2024-09-05T03:29:37.743Z"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "example1": {
                                        "summary": "invalid field",
                                        "value": {
                                            "error": {
                                                "message": "One or more fields are missing or invalid",
                                                "code": "bad.request",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    },
                                    "example2": {
                                        "summary": "invalid zipCode format",
                                        "value": {
                                            "error": {
                                                "message": "invalid zipcode format",
                                                "code": "influencer-service.createInfluencer",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    },
                                    "example3": {
                                        "summary": "error fetching address",
                                        "value": {
                                            "error": {
                                                "message": "error fetching address from ViaCEP",
                                                "code": "create-influencer.getAddress",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": {
                                        "message": "Unauthorized",
                                        "code": "auth-guard.jwtAuth",
                                        "status": true
                                    },
                                    "data": {}
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "example1": {
                                        "summary": "field already in use by another user",
                                        "value": {
                                            "error": {
                                                "message": "['email', 'username'] already taken",
                                                "code": "influencer-repository.createInfluencer",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Error"
                            },
            
                            "examples": {
                                "example1": {
                                    "summary": "service error",
                                    "value": {
                                        "error": {
                                            "message": "failed to create influencer",
                                            "code": "influencer-service.createInfluencer",
                                            "status": true
                                        },
                                        "data": {}
                                    }
                                },
                                "example2": {
                                    "summary": "repository error",
                                    "value": {
                                            "error": {
                                                "message": "influencer not created",
                                                "code": "influencer-repository.createInfluencer",
                                                "status": true
                                                },
                                                "data": {}
                                            }
                                        }
                                    }
                                }
                            }
                    }
                }
            }
        },
        "/influencer/all": {
            "get": {
                "summary": "Get all influencers",
                "description": "Route to find all influencers",
                "tags": ["Influencer"],
                "security": [
                    {
                    "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Bearer {accesToken}",
                    "schema": {
                        "type": "string"
                    }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InfluencersArray"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Ygona Moura",
                                        "username": "ygona.moura",
                                        "email": "ygona@email.com",
                                        "reach": 125000,
                                        "photo": "https://example.photo.aws.com",
                                        "createdAt": "2024-09-05T03:29:37.743Z",
                                        "updatedAt": "2024-09-05T03:29:37.743Z"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": {
                                        "message": "Unauthorized",
                                        "code": "auth-guard.jwtAuth",
                                        "status": true
                                    },
                                    "data": {}
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
            
                                "examples": {
                                    "example1": {
                                        "summary": "service error",
                                        "value": {
                                            "error": {
                                                "message": "failed to get influencers",
                                                "code": "influencer-service.findAllInfluencer",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    },
                                    "example2": {
                                        "summary": "repository error",
                                        "value": {
                                            "error": {
                                                "message": "could not get influencers",
                                                "code": "influencer-repository.findAllInfluencers",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    } 
                                }
                            }
                        }
                    }
                }
            }
        },
        "/influencer/filter": {
            "get": {
                "summary": "Get influencers by filter",
                "description": "Route to find influencers by filter",
                "tags": ["Influencer"],
                "security": [
                    {
                    "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "description": "Bearer {accesToken}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reachMin",
                        "in": "query",
                        "description": "minimum reach",
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "reachMax",
                        "in": "query",
                        "description": "maximun reach",
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "influencer city",
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "niche",
                        "in": "query",
                        "description": "influencer category",
                        "schema": {
                          "type": "array",
                          "items": {
                            "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InfluencerInterfaceArray"
                                },
                                "example": [
                                    {
                                        "id": "1",
                                        "name": "Ygona Moura",
                                        "username": "ygona.moura",
                                        "email": "ygona@email.com",
                                        "niches": ["Artes", "Beleza & Maquiagem"],
                                        "reach": "125000",
                                        "photo": "https://example.photo.aws.com",
                                        "address": {
                                            "street": "Ipê",
                                            "number": "45",
                                            "zipCode": "12345678",
                                            "city": "Belo Horizonte",
                                            "state": "MG"
                                        },
                                        "createdAt": "2024-09-05T03:29:37.743Z",
                                        "updatedAt": "2024-09-05T03:29:37.743Z"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": {
                                        "message": "Unauthorized",
                                        "code": "auth-guard.jwtAuth",
                                        "status": true
                                    },
                                    "data": {}
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
            
                                "examples": {
                                    "example1": {
                                        "summary": "service error",
                                        "value": {
                                            "error": {
                                                "message": "failed to get influencers",
                                                "code": "influencer-service.findInfluencersByFilter",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    },
                                    "example2": {
                                        "summary": "repository error",
                                        "value": {
                                            "error": {
                                                "message": "could not get influencers",
                                                "code": "influencer-repository.findInfluencersByFilter",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    } 
                                }
                            }
                        }
                    }
                }
            }
        },
        "/influencer/{influencerId}": {
            "get": {
                "summary": "Find influencer by id",
                "description": "Route to find influencer by id.",
                "tags": ["Influencer"],
                "security": [
                    {
                    "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "description": "Bearer {accesToken}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "influencer id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer data",
                        "content": {
                            "application/json": {
                                "schema": {
                                "$ref": "#/components/schemas/InfluencerInterface"
                                },
                                "example": {
                                    "id": "1",
                                    "name": "Ygona Moura",
                                    "username": "ygona",
                                    "email": "ygona@email.com",
                                    "niches": ["Beleza & Maquiagem", "Lifestyle"],
                                    "reach": "245987",
                                    "photo": "https://new-example.photo.aws.com",
                                    "address": {
                                        "street": "Ipê",
                                        "number": "45",
                                        "zipCode": "12345678",
                                        "city": "Belo Horizonte",
                                        "state": "MG"
                                    },
                                    "createdAt": "2024-09-05T03:29:37.743Z",
                                    "updatedAt": "2024-09-05T03:29:37.743Z"
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": {
                                        "message": "Unauthorized",
                                        "code": "auth-guard.jwtAuth",
                                        "status": true
                                    },
                                    "data": {}
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Error"
                            },
            
                            "examples": {
                                "example1": {
                                    "summary": "service error",
                                    "value": {
                                        "error": {
                                            "message": "failed to get influencer data",
                                            "code": "influencer-service.findOneInfluencer",
                                            "status": true
                                        },
                                        "data": {}
                                    }
                                },
                                "example2": {
                                    "summary": "repository error",
                                    "value": {
                                            "error": {
                                                "message": "could not get influencer details",
                                                "code": "influencer-repository.findOneInfluencer",
                                                "status": true
                                                },
                                                "data": {}
                                            }
                                        }
                                    }
                                }
                            }
                    }
                }
            }
        },
...
        "influencer/delete/{id}": {
            "delete": {
                "summary": "Delete influencer",
                "description": "Route to delete influencer.",
                "tags": ["Influencer"],
                "security": [
                    {
                    "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "description": "Bearer {accesToken}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "influencerId",
                        "in": "path",
                        "description": "influencer id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": {
                                        "message": "Unauthorized",
                                        "code": "auth-guard.jwtAuth",
                                        "status": true
                                    },
                                    "data": {}
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                
                                "examples": {
                                    "example1": {
                                        "summary": "service error",
                                        "value": {
                                            "error": {
                                                "message": "failed to delete influencer",
                                                "code": "influencer-service.deleteInfluencer",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    },
                                    "example2": {
                                        "summary": "repository error",
                                        "value": {
                                            "error": {
                                                "message": "could not delete influencer",
                                                "code": "influencer-repository.deleteInfluencer",
                                                "status": true
                                            },
                                            "data": {}
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateInfluencer": {
                "allOf": [
                    {"$ref": "#/components/schemas/CreateInfluencerBase"},
                    {
                        "required": [
                            "name",
                            "username",
                            "email",
                            "niches",
                            "reach",
                            "zipCode",
                            "street",
                            "number"
                        ]
                    }
                ]
            },
...
            "InfluencerInterfaceArray": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/InfluencerInterface"
                }
            },
...
            "InfluencersArray": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Influencer"
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "description": "Error message"
                            },
                            "code": {
                                "type": "string",
                                "description": "Error code"
                            },
                            "status": {
                                "type": "boolean",
                                "description": "Error status"
                            }
                        }
                    },
                    "data": {
                        "type": "object",
                        "properties": {}
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}